refactor: address pr review feedback
- replace useEffect+useRef autofocus with callback ref in fees-dropdown
- extract bridge logic from BridgePage into useBridge custom hook
- fix formatting in useBridgeOutbound
